provider_service/requirements.txt
flask
requests
pyjwt

